+++ GeneralsMD/Code/GameEngine/Include/GameNetwork/NetPacket.h
@@ -104,6 +104,7 @@ protected:
 	static UnsignedInt GetDisconnectFrameCommandSize(NetCommandMsg *msg);
 	static UnsignedInt GetDisconnectScreenOffCommandSize(NetCommandMsg *msg);
 	static UnsignedInt GetFrameResendRequestCommandSize(NetCommandMsg *msg);
+	static UnsignedInt GetAllyCursorCommandSize(NetCommandMsg *msg);
 
 	static void FillBufferWithGameCommand(UnsignedByte *buffer, NetCommandRef *msg);
 	static void FillBufferWithAckCommand(UnsignedByte *buffer, NetCommandRef *msg);
@@ -129,6 +130,7 @@ protected:
 	static void FillBufferWithDisconnectFrameMessage(UnsignedByte *buffer, NetCommandRef *msg);
 	static void FillBufferWithDisconnectScreenOffMessage(UnsignedByte *buffer, NetCommandRef *msg);
 	static void FillBufferWithFrameResendRequestMessage(UnsignedByte *buffer, NetCommandRef *msg);
+	static void FillBufferWithAllyCursorMessage(UnsignedByte *buffer, NetCommandRef *msg);
 
 	Bool addFrameCommand(NetCommandRef *msg);
 	Bool isRoomForFrameMessage(NetCommandRef *msg);
@@ -183,6 +185,8 @@ protected:
 	Bool isRoomForDisconnectScreenOffMessage(NetCommandRef *msg);
 	Bool addFrameResendRequestCommand(NetCommandRef *msg);
 	Bool isRoomForFrameResendRequestMessage(NetCommandRef *msg);
+	Bool addAllyCursorCommand(NetCommandRef *msg);
+	Bool isRoomForAllyCursorMessage(NetCommandRef *msg);
 
 	Bool isAckRepeat(NetCommandRef *msg);
 	Bool isAckBothRepeat(NetCommandRef *msg);
@@ -217,6 +221,7 @@ protected:
 	static NetCommandMsg * readDisconnectFrameMessage(UnsignedByte *data, Int &i);
 	static NetCommandMsg * readDisconnectScreenOffMessage(UnsignedByte *data, Int &i);
 	static NetCommandMsg * readFrameResendRequestMessage(UnsignedByte *data, Int &i);
+	static NetCommandMsg * readAllyCursorMessage(UnsignedByte *data, Int &i);
 
 	void writeGameMessageArgumentToPacket(GameMessageArgumentDataType type, GameMessageArgumentType arg);
 	static void readGameMessageArgumentFromPacket(GameMessageArgumentDataType type, NetGameCommandMsg *msg, UnsignedByte *data, Int &i);