+++ GeneralsMD/Code/GameEngine/Source/GameNetwork/NetCommandMsg.cpp
@@ -763,6 +763,36 @@ void NetChatCommandMsg::setPlayerMask( Int playerMask )
 }
 
 //-------------------------
+// NetAllyCursorCommandMsg
+//-------------------------
+NetAllyCursorCommandMsg::NetAllyCursorCommandMsg() : NetCommandMsg()
+{
+	m_commandType = NETCOMMANDTYPE_ALLYCURSOR;
+	m_x = 0.0f;
+	m_y = 0.0f;
+}
+
+NetAllyCursorCommandMsg::~NetAllyCursorCommandMsg()
+{
+}
+
+Real NetAllyCursorCommandMsg::getX()
+{
+	return m_x;
+}
+
+Real NetAllyCursorCommandMsg::getY()
+{
+	return m_y;
+}
+
+void NetAllyCursorCommandMsg::setPosition( Real x, Real y )
+{
+	m_x = x;
+	m_y = y;
+}
+
+//-------------------------
 // NetDisconnectVoteCommandMsg
 //-------------------------
 /**