+++ GeneralsMD/Code/GameEngine/Source/GameNetwork/NetworkUtil.cpp
@@ -294,6 +294,8 @@ AsciiString GetAsciiNetCommandType(NetCommandType type) {
 		s.set("NETCOMMANDTYPE_DISCONNECTSCREENOFF");
 	} else if (type == NETCOMMANDTYPE_FRAMERESENDREQUEST) {
 		s.set("NETCOMMANDTYPE_FRAMERESENDREQUEST");
+	} else if (type == NETCOMMANDTYPE_ALLYCURSOR) {
+		s.set("NETCOMMANDTYPE_ALLYCURSOR");
 	} else {
 		s.set("UNKNOWN");
 	}